+++ template/TypeDossierPutInFatalError.php
@@ -2,6 +2,7 @@
 
 /** @var Gabarit $this */
 /** @var array $type_de_dossier_info */
+/** @var String $content */
 
 ?>
 
@@ -10,7 +11,10 @@
     <div class="alert-danger alert">
         Vous êtes sur le point de mettre tous les dossiers associés au flux <b><?php hecho($type_de_dossier_info['id_type_dossier']) ?></b> en erreur fatale.<br>
         Attention ! Cette action est <b>définitive</b>, vous ne pourrez pas revenir en arrière.</div>
-    <form action='<?php $this->url("/TypeDossier/doSetFile"); ?>' method='post' >
+
+    <div><?php echo $content; ?></div>
+
+    <form action='<?php $this->url('/TypeDossier/doPutInFatalError'); ?>' method='post' >
         <?php $this->displayCSRFInput() ?>
         <input type='hidden' name='id_t' value='<?php hecho($type_de_dossier_info['id_t'])?>' />
         <input type='hidden' name='id_type_dossier' value='<?php hecho($type_de_dossier_info['id_type_dossier'])?>' />
@@ -19,7 +23,7 @@
             <i class="fa fa-times-circle"></i>&nbsp;Annuler
         </a>
         <button type="submit" class="btn btn-danger">
-            <i class="fa fa-folder"></i>&nbsp;Mettre tous les dossiers en 'Fatal error'
+            <i class="fa fa-folder"></i>&nbsp;Mettre tous les dossiers en erreur fatale
         </button>
 
     </form>